Changes to support webpack
Since webpack minifies and uglifies classes, we can't have code dependent on constructor names.  Fix this via special getters.

In addition, fix the package.json which was erroneously pointing to at the (now) non-existent lib folder.

Note that there is still one thing that fails webpack, and that is the ucum validation in quantity.js.  It *always* returns false when webpacked.